refactor MAC db into a DBM file for persitence across startup
Previous behavior was to load the MAC addrs from db/*.txt at startup.
The DB's are fairly small and the performance impact is neglible. Still
it is preferable to parse the txt files just once and read the parsed
format in at startup.

Implemented with ruby stdlib dbm library. Future plans are to convert to
sqlite.